created folder structure for project
created shell scripts to automate creating database and starting server
implemented declarative base for sql alchemy
separated services, database and modes into separate files
updated instructions.txt
created requirements.txt
