+++ database/create_db.sh
@@ -1,0 +1,2 @@
+rm database/development.db
+touch database/development.db
