+++ giftapp/database/database.py
@@ -1,0 +1,20 @@
+from configparser import ConfigParser
+from sqlalchemy import create_engine
+from sqlalchemy.orm import scoped_session, sessionmaker
+from sqlalchemy.ext.declarative import declarative_base
+
+config = ConfigParser()
+config.read('config/configuration.conf')
+db_options = dict(config['DATABASE'])
+engine = create_engine(db_options['sqlite_connection_string'])
+db_session = scoped_session(sessionmaker(autocommit=False,
+                                         autoflush=False,
+                                         bind=engine))
+Base = declarative_base()
+Base.query = db_session.query_property()
+
+def init_db():
+    # import all modules here that might define models so that
+    # they will be registered properly on the metadata.  Otherwise
+    # you will have to import them first before calling init_db()
+    Base.metadata.create_all(bind=engine)
