+++ requirements.txt
@@ -1,0 +1,9 @@
+click==8.0.4
+Flask==2.0.3
+Flask-SQLAlchemy==2.5.1
+greenlet==1.1.2
+itsdangerous==2.1.1
+Jinja2==3.0.3
+MarkupSafe==2.1.1
+SQLAlchemy==1.4.32
+Werkzeug==2.0.3
